Add script to change cursor per mode
The cursor will become a vertical line in insert mode and an underscore
in replace mode.

This is confirmed to work fully in VTE-compatible terminals outside of
TMUX in both Vim and Neovim.  Vim's support extends to running TMUX on
top of the VTE terminal, but the current configuration under Neovim does
nothing when inside TMUX.