[nvim] Fix `StripEndWhitespace()`
`StripEndWhitespace()` is renamed `StripEndLine()`.  The autocmd in
autocmd.vim has been updated to call the new function name.

`StripEndLine()` uses a top-down regex.  This approach likely has
decreased performance relative to the bottom-up deletion of
`StripEndWhitespace()`.  Neither function's performance was ever
measured, so it is unknown exactly what the decrease in performance is.

`StripEndLine()` uses two per-buffer variables.  Defining
`b:keep_end_lines` causes no action to be performed (i.e., does not
remove any lines).  `b:strip_end_regex` can be set to a regex string to
match the desired lines to remove;  the default value if it is left
undefined is '\s*' which matches blank and empty lines.