[emacs][init-window-and-frame] Enable Ace Window for EXWM buffers
The Ace Window key for a given window is displayed in the window's modeline.
The modelines of Firefox or floating windows are displayed so all EXWM buffers
will be labeled with their Ace Window keys.  The keychord used for ~ace-window~
in the global keymap is added as an EXWM input.