[emacs][init-eshell] Cleanup Org file
Only a single source block is tangled.  The rest of the source is included in
that block through hierarchical noweb-refs.  This simplifies all other source
block headers at the cost of less granularity in what gets tangled; all source
blocks under the Code header are written in the tangle output.